test(all): rename documents
