+++ tests/Functional/CertificateTest.php
@@ -14,7 +14,7 @@ namespace Unikorp\KongAdminApi\Tests\Functional;
 use PHPUnit\Framework\TestCase;
 use Unikorp\KongAdminApi\Client;
 use Unikorp\KongAdminApi\Configurator;
-use Unikorp\KongAdminApi\Document\Certificate as Document;
+use Unikorp\KongAdminApi\Document\CertificateDocument as Document;
 use Unikorp\KongAdminApi\Node\Certificate as Node;
 
 /**
