+++ tests/Unit/Document/InformationDocumentTest.php
@@ -11,19 +11,19 @@
 
 namespace Unikorp\KongAdminApi\Tests\Unit\Document;
 
-use Unikorp\KongAdminApi\Document\Consumer as Document;
+use Unikorp\KongAdminApi\Document\InformationDocument as Document;
 use PHPUnit\Framework\TestCase;
 
 /**
- * consumer test
+ * information document test
  *
  * @author VEBER Arnaud <https://github.com/VEBERArnaud>
  */
-class ConsumerTest extends TestCase
+class InformationDocumentTest extends TestCase
 {
     /**
      * document
-     * @var \Unikorp\KongAdminApi\Document\Consumer $document
+     * @var \Unikorp\KongAdminApi\Document\InformationDocument $document
      */
     private $document = null;
 
@@ -49,83 +49,6 @@ class ConsumerTest extends TestCase
     protected function tearDown()
     {
         $this->document = null;
-    }
-
-    /**
-     * test set username
-     *
-     * @return void
-     *
-     * @covers \Unikorp\KongAdminApi\Document\Consumer::setUsername
-     */
-    public function testSetUsername()
-    {
-        // asserts
-        $this->document->setUsername('test');
-        $this->assertSame('test', $this->readAttribute($this->document, 'username'));
-    }
-
-    /**
-     * test get username
-     *
-     * @return void
-     *
-     * @covers \Unikorp\KongAdminApi\Document\Consumer::getUsername
-     */
-    public function testGetUsername()
-    {
-        // reflect `document`
-        $reflectionClass = new \ReflectionClass($this->document);
-
-        // set `username` property from `document` accessible
-        $reflectionProperty = $reflectionClass->getProperty('username');
-        $reflectionProperty->setAccessible(true);
-
-        // assert
-        $reflectionProperty->setValue($this->document, 'test');
-        $this->assertSame('test', $this->document->getUsername());
-    }
-
-    /**
-     * test set custom id
-     *
-     * @return void
-     *
-     * @covers \Unikorp\KongAdminApi\Document\Consumer::setCustomId
-     */
-    public function testCustomId()
-    {
-        // reflect `document`
-        $reflectionClass = new \ReflectionClass($this->document);
-
-        // set `custom id` property from `document` accessible
-        $reflectionProperty = $reflectionClass->getProperty('customId');
-        $reflectionProperty->setAccessible(true);
-
-        // asserts
-        $this->document->setCustomId('test');
-        $this->assertSame('test', $reflectionProperty->getValue($this->document));
-    }
-
-    /**
-     * test get custom id
-     *
-     * @return void
-     *
-     * @covers \Unikorp\KongAdminApi\Document\Consumer::getCustomId
-     */
-    public function testGetCustomId()
-    {
-        // reflect `document`
-        $reflectionClass = new \ReflectionClass($this->document);
-
-        // set `custom id` property from `document` accessible
-        $reflectionProperty = $reflectionClass->getProperty('customId');
-        $reflectionProperty->setAccessible(true);
-
-        // assert
-        $reflectionProperty->setValue($this->document, 'test');
-        $this->assertSame('test', $this->document->getCustomId());
     }
 
     /**
@@ -241,20 +164,17 @@ class ConsumerTest extends TestCase
      *
      * @covers \Unikorp\KongAdminApi\AbstractDocument::toJson
      * @covers \Unikorp\KongAdminApi\AbstractDocument::toRequestParameters
-     * @covers \Unikorp\KongAdminApi\Document\Consumer::getFields
-     * @covers \Unikorp\KongAdminApi\AbstractDocument::toSnakeCase
+     * @covers \Unikorp\KongAdminApi\Document\InformationDocument::getFields
      */
     public function testToJson()
     {
         $this->document
-            ->setUsername('username')
-            ->setCustomId('customId')
             ->setCreatedAt(42)
             ->setSize(50)
             ->setOffset('offset');
 
         $this->assertSame(
-            '{"username":"username","custom_id":"customId","created_at":42,"size":50,"offset":"offset"}',
+            '{"created_at":42,"size":50,"offset":"offset"}',
             $this->document->toJson()
         );
     }
@@ -266,20 +186,17 @@ class ConsumerTest extends TestCase
      *
      * @covers \Unikorp\KongAdminApi\AbstractDocument::toQueryString
      * @covers \Unikorp\KongAdminApi\AbstractDocument::toRequestParameters
-     * @covers \Unikorp\KongAdminApi\Document\Consumer::getFields
-     * @covers \Unikorp\KongAdminApi\AbstractDocument::toSnakeCase
+     * @covers \Unikorp\KongAdminApi\Document\InformationDocument::getFields
      */
     public function testToQueryString()
     {
         $this->document
-            ->setUsername('username')
-            ->setCustomId('customId')
             ->setCreatedAt(42)
             ->setSize(50)
             ->setOffset('offset');
 
         $this->assertSame(
-            'username=username&custom_id=customId&created_at=42&size=50&offset=offset',
+            'created_at=42&size=50&offset=offset',
             $this->document->toQueryString()
         );
     }
