+++ tests/Unit/Node/ClusterTest.php
@@ -11,6 +11,7 @@
 
 namespace Unikorp\KongAdminApi\Tests\Unit\Node;
 
+use Unikorp\KongAdminApi\Document\ClusterDocument as Document;
 use Unikorp\KongAdminApi\Node\Cluster as Node;
 use PHPUnit\Framework\TestCase;
 
@@ -128,7 +129,7 @@ class ClusterTest extends TestCase
             ->will($this->returnValue($this->httpClient));
 
         // mock `document`
-        $document = $this->createMock('\Unikorp\KongAdminApi\Document\Cluster');
+        $document = $this->createMock(Document::class);
 
         // mock `response`
         $response = $this->createMock('\GuzzleHttp\Psr7\Response');
@@ -168,7 +169,7 @@ class ClusterTest extends TestCase
             ->will($this->returnValue($this->httpClient));
 
         // mock `document`
-        $document = $this->createMock('\Unikorp\KongAdminApi\Document\Cluster');
+        $document = $this->createMock(Document::class);
 
         // mock `response`
         $response = $this->createMock('\GuzzleHttp\Psr7\Response');
