+++ tests/Unit/Node/TargetTest.php
@@ -11,6 +11,7 @@
 
 namespace Unikorp\KongAdminApi\Tests\Unit\Node;
 
+use Unikorp\KongAdminApi\Document\TargetDocument as Document;
 use Unikorp\KongAdminApi\Node\Target as Node;
 use PHPUnit\Framework\TestCase;
 
@@ -97,7 +98,7 @@ class TargetTest extends TestCase
             ->will($this->returnValue($this->httpClient));
 
         // mock `document`
-        $document = $this->createMock('\Unikorp\KongAdminApi\Document\Target');
+        $document = $this->createMock(Document::class);
 
         // mock `response`
         $response = $this->createMock('\GuzzleHttp\Psr7\Response');
